Move PrimitiveFactory to its own File, Rename to just Primitives. Add Overlay.Primitive Extension. Move Shaders to the Shaders folder
