--- Runtime/Utility/Debug/Overlay/DebugGeo.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 03906f5c75f01324690f89166dbebe4c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
