--- Runtime/Utility/Debug/Overlay/DebugOverlay.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e17f6c7aac9a98a4088e45638d5c5a85
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
